+++ aquadoggo/src/db/stores/mod.rs
@@ -6,7 +6,6 @@ pub mod document;
 mod entry;
 mod log;
 mod operation;
-#[allow(unused)]
 mod query;
 mod schema;
 mod task;