+++ CHANGELOG.md
@@ -42,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Handle decoding and parsing empty pinned relation lists [#336](https://github.com/p2panda/aquadoggo/pull/336)
 - Make `value` on `QueryRow` an option [#339](https://github.com/p2panda/aquadoggo/pull/339)
 - Fix race conditions in SQLite database test runner [#345](https://github.com/p2panda/aquadoggo/pull/345)
+- Remove duplicate `Cursor` implementation, query code clean up [#346](https://github.com/p2panda/aquadoggo/pull/346)
 - Fix SQL ordering to correctly assemble resulting rows to documents [#347](https://github.com/p2panda/aquadoggo/pull/347)
 
 ## [0.4.0]